requirements.txt
# Core Web Framework
django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0

# Database & Infrastructure
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
dj-database-url>=2.1.0
gunicorn>=21.2.0

# Authentication (OAuth2 & JWT)
# Required for access + refresh tokens and session management [cite: 44, 112-115]
djangorestframework-simplejwt>=5.3.0
requests-oauthlib>=1.3.1
requests>=2.31.0

# Querying & Search
# Powering advanced filtering and Postgres Full-Text Search [cite: 9, 17]
django-filter>=23.2

# Utilities
# Used for UUID v7 generation and CSV handling [cite: 118, 203]
uuid6>=2024.1.12